+++ Demos/AllOverIt.Csv/CsvExporterDemo/Data/Coordinates.cs
@@ -1,4 +1,4 @@
-﻿namespace CsvExportDemo.Data
+﻿namespace CsvExporterDemo.Data
 {
     public sealed class Coordinates
     {