+++ Demos/AllOverIt.Csv/CsvSerializerDemo/Data/MetadataType.cs
@@ -1,0 +1,8 @@
+﻿namespace CsvSerializerDemo.Data
+{
+    public enum MetadataType
+    {
+        Environment,
+        Quality
+    }
+}