+++ Demos/AllOverIt.Csv/CsvSerializerDemo/Data/SampleMetadata.cs
@@ -1,4 +1,4 @@
-﻿namespace CsvExportDemo.Data
+﻿namespace CsvSerializerDemo.Data
 {
     public sealed class SampleMetadata
     {