+++ Source/AllOverIt.Csv/Exporter/BufferedCsvExporterBase.cs
@@ -5,58 +5,35 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace AllOverIt.Csv
+namespace AllOverIt.Csv.Exporter
 {
-    public interface IBufferedCsvSerializer
-        <TModel> : IAsyncDisposable
+    public abstract class BufferedCsvExporterBase<TModel> : IBufferedCsvExporter<TModel>
     {
-        void Configure(IEnumerable<TModel> dynamicFields = default);
-        Task AddDataAsync(TModel data, CancellationToken cancellationToken);
-        Task<ReadOnlyMemory<byte>> GetContentAsync(CancellationToken cancellationToken);
-    }
-
-
-
-    public sealed class BufferedCsvSerializerConfiguration
-    {
-        public int BufferSize { get; init; } = 16;
-        public bool IncludeHeaders { get; init; } = true;
-    }
-
-
-
-    public abstract class BufferedCsvSerializerBase<TModel> : IBufferedCsvSerializer<TModel>
-    {
-        private readonly BufferedCsvSerializerConfiguration _configuration;
+        private readonly BufferedCsvExporterConfiguration _configuration;
         private readonly IList<TModel> _data;
 
         private StreamWriter _writer;
         private ICsvSerializer<TModel> _csvSerializer;
 
-        protected BufferedCsvSerializerBase()
-            : this(new BufferedCsvSerializerConfiguration())
+        protected Stream Stream { get; private set; }       // The underlying stream passed to the StreamWriter
+
+        protected BufferedCsvExporterBase()
+            : this(new BufferedCsvExporterConfiguration())
         {
         }
 
-        protected BufferedCsvSerializerBase(BufferedCsvSerializerConfiguration configuration)
+        protected BufferedCsvExporterBase(BufferedCsvExporterConfiguration configuration)
         {
             _configuration = configuration.WhenNotNull(nameof(configuration));
 
             _data = new List<TModel>(_configuration.BufferSize);
         }
 
-        public async ValueTask DisposeAsync()
-        {
-            await DisposeAsyncCore().ConfigureAwait(false);
-
-            GC.SuppressFinalize(this);
-        }
-
-        public void Configure(IEnumerable<TModel> dynamicFields = default)
+        public void Configure(IEnumerable<TModel> dynamicData = default)
         {
             Throw<InvalidOperationException>.WhenNotNull(_csvSerializer, "The CSV serializer is already configured.");
 
-            _csvSerializer = CreateSerializer(dynamicFields);
+            _csvSerializer = CreateSerializer(dynamicData);
         }
 
         public Task AddDataAsync(TModel data, CancellationToken cancellationToken)
@@ -66,21 +43,30 @@ namespace AllOverIt.Csv
             return ProcessBufferAsync(false, cancellationToken);
         }
 
-        public async Task<ReadOnlyMemory<byte>> GetContentAsync(CancellationToken cancellationToken)
+        public async ValueTask DisposeAsync()
+        {
+            await DisposeAsyncCore().ConfigureAwait(false);
+
+            GC.SuppressFinalize(this);
+        }
+
+        protected abstract Stream CreateStream();
+
+        protected abstract ICsvSerializer<TModel> CreateSerializer(IEnumerable<TModel> dynamicData = default);
+
+        protected async Task FlushAsync(CancellationToken cancellationToken)
         {
             await ProcessBufferAsync(true, cancellationToken);
 
             await _writer.FlushAsync();
-
-            return ((MemoryStream) _writer.BaseStream).ToArray();
         }
-
-        protected abstract ICsvSerializer<TModel> CreateSerializer(IEnumerable<TModel> dynamicFields = default);
 
         protected virtual async ValueTask DisposeAsyncCore()
         {
             if (_writer is not null)
             {
+                await FlushAsync(CancellationToken.None);
+
                 await _writer.DisposeAsync();
                 _writer = null;
             }
@@ -94,7 +80,12 @@ namespace AllOverIt.Csv
             {
                 var includeHeaders = _configuration.IncludeHeaders && _writer is null;
 
-                _writer ??= new(new MemoryStream(), leaveOpen: false);
+                if (_writer is null)
+                {
+                    Stream = CreateStream();
+
+                    _writer = new StreamWriter(Stream, null, -1, false);
+                }
 
                 await _csvSerializer.SerializeAsync(_writer, _data, includeHeaders, true, cancellationToken);