+++ Source/AllOverIt.Csv/Exporter/IMemoryCsvExporter.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace AllOverIt.Csv.Exporter
+{
+    public interface IMemoryCsvExporter<TModel> : IBufferedCsvExporter<TModel>
+    {
+        Task<byte[]> GetContentAsync(CancellationToken cancellationToken);
+    }
+}